MNT: Rename modules (xtcedef / csvdev -> definitions)
