+++ examples/parsing_and_plotting_idex_waveforms_from_socket.py
@@ -19,7 +19,7 @@ import time
 # Installed
 import matplotlib.pyplot as plt
 # Local
-from space_packet_parser import xtcedef
+from space_packet_parser import definitions
 from space_packet_parser import parser
 
 
@@ -99,8 +99,8 @@ if __name__ == "__main__":
     """Parse IDEX data"""
     idex_test_data_dir = Path("../tests/test_data/idex")
     idex_xtce = idex_test_data_dir / 'idex_combined_science_definition.xml'
-    idex_definition = xtcedef.XtcePacketDefinition(xtce_document=idex_xtce)
-    assert isinstance(idex_definition, xtcedef.XtcePacketDefinition)
+    idex_definition = definitions.XtcePacketDefinition(xtce_document=idex_xtce)
+    assert isinstance(idex_definition, definitions.XtcePacketDefinition)
     idex_parser = parser.PacketParser(idex_definition)
     idex_packet_file = idex_test_data_dir / 'sciData_2023_052_14_45_05'
