+++ tests/integration/test_bufferedreader_parsing.py
@@ -1,14 +1,14 @@
 """Integration test for parsing JPSS packets"""
 # Local
-from space_packet_parser import xtcedef
+from space_packet_parser import definitions
 from space_packet_parser import parser, parseables
 
 
 def test_jpss_xtce_packet_parsing(jpss_test_data_dir):
     """Test parsing a real XTCE document"""
     jpss_xtce = jpss_test_data_dir / 'jpss1_geolocation_xtce_v1.xml'
-    jpss_definition = xtcedef.XtcePacketDefinition(xtce_document=jpss_xtce)
-    assert isinstance(jpss_definition, xtcedef.XtcePacketDefinition)
+    jpss_definition = definitions.XtcePacketDefinition(xtce_document=jpss_xtce)
+    assert isinstance(jpss_definition, definitions.XtcePacketDefinition)
     jpss_parser = parser.PacketParser(jpss_definition)
 
     jpss_packet_file = jpss_test_data_dir / 'J01_G011_LZ_2021-04-09T00-00-00Z_V01.DAT1'
