+++ tests/integration/test_xtce_based_parsing/test_inheritance_restrictions.py
@@ -1,14 +1,14 @@
 """Test RestrictionCriteria being used creatively with JPSS data"""
 # Local
-from space_packet_parser import xtcedef
+from space_packet_parser import definitions
 from space_packet_parser import parser, parseables
 
 
 def test_jpss_xtce_packet_parsing(jpss_test_data_dir):
     """Test parsing a real XTCE document"""
     jpss_xtce = jpss_test_data_dir / 'contrived_inheritance_structure.xml'
-    jpss_definition = xtcedef.XtcePacketDefinition(xtce_document=jpss_xtce)
-    assert isinstance(jpss_definition, xtcedef.XtcePacketDefinition)
+    jpss_definition = definitions.XtcePacketDefinition(xtce_document=jpss_xtce)
+    assert isinstance(jpss_definition, definitions.XtcePacketDefinition)
     jpss_parser = parser.PacketParser(jpss_definition)
 
     jpss_packet_file = jpss_test_data_dir / 'J01_G011_LZ_2021-04-09T00-00-00Z_V01.DAT1'
