+++ tests/unit/test_xtcedef.py
@@ -6,7 +6,7 @@ import pytest
 import lxml.etree as ElementTree
 # Local
 from space_packet_parser.exceptions import CalibrationError, ComparisonError
-from space_packet_parser import calibrators, comparisons, encodings, parameters, parseables, parser, xtcedef
+from space_packet_parser import calibrators, comparisons, definitions, encodings, parameters, parseables, parser
 
 XTCE_URI = "http://www.omg.org/space/xtce"
 TEST_NAMESPACE = {'xtce': XTCE_URI}
@@ -41,8 +41,8 @@ def test_invalid_parameter_type_error(test_data_dir):
 </xtce:SpaceSystem>
 """
     x = io.TextIOWrapper(io.BytesIO(test_xtce_document.encode("UTF-8")))
-    with pytest.raises(xtcedef.InvalidParameterTypeError):
-        xtcedef.XtcePacketDefinition(x)
+    with pytest.raises(definitions.InvalidParameterTypeError):
+        definitions.XtcePacketDefinition(x)
 
 
 def test_unsupported_parameter_type_error(test_data_dir):
@@ -83,7 +83,7 @@ def test_unsupported_parameter_type_error(test_data_dir):
 """
     x = io.TextIOWrapper(io.BytesIO(test_xtce_document.encode("UTF-8")))
     with pytest.raises(NotImplementedError):
-        xtcedef.XtcePacketDefinition(x)
+        definitions.XtcePacketDefinition(x)
 
 
 def test_attr_comparable():
@@ -809,7 +809,7 @@ def test_polynomial_calibrator_calibrate(xq, expectation):
     </xtce:SizeInBits>
 </xtce:StringDataEncoding>
 """,
-         xtcedef.ElementNotFoundError())
+         definitions.ElementNotFoundError())
         ]
 )
 def test_string_data_encoding(xml_string: str, expectation):
@@ -2033,7 +2033,7 @@ def test_parameter():
 def test_parsing_xtce_document(test_data_dir):
     """Tests parsing an entire XTCE document and makes assertions about the contents"""
     with open(test_data_dir / "test_xtce.xml") as x:
-        xdef = xtcedef.XtcePacketDefinition(x, ns=TEST_NAMESPACE)
+        xdef = definitions.XtcePacketDefinition(x, ns=TEST_NAMESPACE)
 
     # Test Parameter Types
     ptname = "USEC_Type"
